Safe working with bytes
String#size methods returns number of UTF-8 endpoints, not number of bytes, also String#[] is designed for UTF-8 strings and will raise exception on raw bytes. But I can store any data on Redis like MessagePack strings or images - so there will be an error in this case. This patch fix it.